pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "zoom-dl"
version = "1.0.0"
description = "Download Zoom cloud recordings from share links"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Aakash", email = "aakash.lohchab97@gmail.com"},
]
keywords = ["zoom", "downloader", "recording", "video"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Utilities",
]
dependencies = [
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
]

[project.optional-dependencies]
browser = ["playwright>=1.40.0"]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
zoom-dl = "zoomdl.cli:main"

[project.urls]
Repository = "https://github.com/DDG0D/zoom-dl"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
